+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 2.0.1
+## 2.1.0
+
+### Added
+
+- New variables {{bookId}} and {{editionId}} available in both filename and body templates.
+
+#### Fixed
+
+- General code cleanup and improvements. No change in functionality.
 
 ### Fixed